feat(skipLast): add higher-order lettable version of skipLast
