+++ src/operators/index.ts
@@ -61,6 +61,7 @@ export { reduce } from './reduce';
 export { refCount } from './refCount';
 export { scan } from './scan';
 export { skip } from './skip';
+export { skipLast } from './skipLast';
 export { subscribeOn } from './subscribeOn';
 export { switchAll } from './switchAll';
 export { switchMap } from './switchMap';
